fix dolls dying even when death was prevented

+++ AveMujicaCode/Cards/Allies/AbstractAlly.cs
@@ -101,7 +101,7 @@ public abstract class AbstractAlly : CustomMonsterModel
     bool wasRemovalPrevented,
     float deathAnimLength)
   {
-    if (creature == Creature.PetOwner?.Creature)
+    if (creature == Creature.PetOwner?.Creature && !wasRemovalPrevented)
     { 
       await CreatureCmd.Kill(Creature);
     }
